+++ libraries/Geometry/src/Cylinder.cpp
@@ -4,19 +4,19 @@
 
 #include "Cylinder.hpp"
 #include "Intersections.hpp"
-#include "FloatUtils.hpp"
+#include "floatUtils.hpp"
 
 namespace raytracer {
 namespace geometry {
 
-static inline bool isIntersectionWithinRadius(const utility::Ray &ray, double t) noexcept {
+static inline bool isIntersectionWithinRadius(const utility::Ray &ray, float t) noexcept {
   auto x = ray.origin_.x() + t * ray.direction_.x();
   auto z = ray.origin_.z() + t * ray.direction_.z();
   return (x * x + z * z) <= 1; // radius is always considered to be 1
 }
 
 static inline std::vector<Intersection> intersectCaps(const Cylinder &cylinder, const utility::Ray &ray) {
-  if (!cylinder.closed_ || utility::floatNearlyEqual(ray.direction_.y(), 0.0)) { return {}; }
+  if (!cylinder.closed_ || utility::floatNearlyEqual(ray.direction_.y(), 0.0f)) { return {}; }
 
   auto intersections = std::vector<Intersection>{};
   auto tLowerCap = (cylinder.minimum_ - ray.origin_.y()) / ray.direction_.y();
@@ -31,7 +31,7 @@ static inline std::vector<Intersection> intersectCaps(const Cylinder &cylinder, 
 std::vector<Intersection> Cylinder::localIntersect(const utility::Ray &transformedRay) const noexcept {
   auto a = transformedRay.direction_.x() * transformedRay.direction_.x() + 
            transformedRay.direction_.z() * transformedRay.direction_.z();
-  if (utility::floatNearlyEqual(a, 0.0)) { return (closed_ ? intersectCaps(*this, transformedRay) : std::vector<Intersection>{}); }
+  if (utility::floatNearlyEqual(a, 0.0f)) { return (closed_ ? intersectCaps(*this, transformedRay) : std::vector<Intersection>{}); }
 
   auto b = 2 * transformedRay.origin_.x() * transformedRay.direction_.x() + 
            2 * transformedRay.origin_.z() * transformedRay.direction_.z();
@@ -47,7 +47,7 @@ std::vector<Intersection> Cylinder::localIntersect(const utility::Ray &transform
   auto t1 = (-b + utility::sqrt(discriminant)) / (2 * a);
 
   auto intersections = std::vector<Intersection>{};
-  auto isBetweenMinAndMax = [this](double y) { return y > minimum_ && y < maximum_; };
+  auto isBetweenMinAndMax = [this](float y) { return y > minimum_ && y < maximum_; };
 
   auto firstIntersectionY = transformedRay.origin_.y() + t0 * transformedRay.direction_.y();
   if (isBetweenMinAndMax(firstIntersectionY)){intersections.emplace_back(Intersection(this, t0));}
@@ -63,8 +63,8 @@ std::vector<Intersection> Cylinder::localIntersect(const utility::Ray &transform
 
 utility::Tuple Cylinder::localNormalAt(const utility::Tuple &objectPoint) const noexcept {
   auto distanceFromYSquared = objectPoint.x() * objectPoint.x() + objectPoint.z() * objectPoint.z();
-  if (distanceFromYSquared < 1 && objectPoint.y() >= maximum_ - utility::EPSILON<double>) { return utility::Vector(0, 1, 0); }
-  else if (distanceFromYSquared < 1 && objectPoint.y() <= minimum_ + utility::EPSILON<double>) { return utility::Vector(0, -1, 0); }
+  if (distanceFromYSquared < 1 && objectPoint.y() >= maximum_ - utility::EPSILON<float>) { return utility::Vector(0, 1, 0); }
+  else if (distanceFromYSquared < 1 && objectPoint.y() <= minimum_ + utility::EPSILON<float>) { return utility::Vector(0, -1, 0); }
   return utility::Vector(objectPoint.x(), 0, objectPoint.z());
 }
